Split block_parser.dart into small files
