+++ lib/src/block_syntaxes/horizontal_rule_syntax.dart
@@ -1,0 +1,18 @@
+import '../ast.dart';
+import '../block_parser.dart';
+import '../patterns.dart';
+import 'block_syntax.dart';
+
+/// Parses horizontal rules like `---`, `_ _ _`, `*  *  *`, etc.
+class HorizontalRuleSyntax extends BlockSyntax {
+  const HorizontalRuleSyntax();
+
+  @override
+  RegExp get pattern => hrPattern;
+
+  @override
+  Node parse(BlockParser parser) {
+    parser.advance();
+    return Element.empty('hr');
+  }
+}
